jedi,vis,report: move units handling to its own submodule
While we're here, also move validate_hessian to the new utils submodule
to improve cleanliness of the main file.
